+++ README.md
@@ -1,2 +1,5 @@
 # Garden
 Golang  launch start samples
+
+- Manual
+  * Please use bash install to start compile and run the application.
