--- http/template/t.htm
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>  
-    <head>
-    <title>Go Template</title>
-    </head>  
-    <body>  
-        <form action="/test" method="POST">  
-        {{with .User}}  
-            {{range .}}  
-            <label for="numbers">Name:</label> 
-            <input type="radio" name="test" value={{.Name}}/>{{.Name}}<br/>
-            <label for="numbers">Age:</label>
-            <input type="text" name="test" value={{.Age}}>{{.Age}}<br/> 
-                {{range .Emails}}
-                <b> {{.}} </b>
-                {{end}}<br/> 
-            {{end}}  
-        {{end}}  
-        <input type="submit" value="submit"/>  
-        </form>  
-    </body>  
-</html>
